♻️ Refactor client cases and initial actions.
