--- src/saragurosnet/cases/__init__.py
@@ -1,5 +1,0 @@
-from src.saragurosnet.cases.client_cases import ClientUseCases
-
-__all__ = [
-    "ClientUseCases",
-]
